delivery location track list

+++ app/Services/DeliveryProductService.php
@@ -2,6 +2,7 @@
 
 namespace App\Services;
 
+use App\Delivery;
 use App\File;
 use App\DeliveryProduct;
 use Illuminate\Support\Facades\Log;
@@ -20,6 +21,12 @@ class DeliveryProductService
 
     public function create($data)
     {
+        $delivery = Delivery::find($data['delivery_id']);
+
+        if (!$delivery->canBeAltered()) {
+            throw new \Exception("El estatus de la entega no permite agregar productos", 1);
+        }
+
         $deliveryProduct = DeliveryProduct::create($data);
 
         if (isset($data['product_image']) && FileService::isBase64Image($data['product_image'])) {
@@ -31,6 +38,10 @@ class DeliveryProductService
 
     public function update(DeliveryProduct $deliveryProduct, $data)
     {
+        if (!$deliveryProduct->delivery->canBeAltered()) {
+            throw new \Exception("El estatus de la entega no permite actualizar productos", 1);
+        }
+
         $deliveryProduct->update($data);
 
         if (isset($data['product_image']) && FileService::isBase64Image($data['product_image'])) {
@@ -38,8 +49,12 @@ class DeliveryProductService
         }
     }
 
-    public function delete(DeliveryProduct $deliveryProduct)
+    public function destroy(DeliveryProduct $deliveryProduct)
     {
+        if (!$deliveryProduct->delivery->canBeAltered()) {
+            throw new \Exception("El estatus de la entega no permite eliminar productos", 1);
+        }
+
         if ($deliveryProduct->productImage()) {
             File::delete_file($deliveryProduct->productImage()->path);
         }
